+++ lib/src/helpers/test_mode/dbas_sqlite_test_mode_io.dart
@@ -1,0 +1,4 @@
+import 'dart:io';
+
+bool isFlutterTestEnv() =>
+    Platform.environment.containsKey('FLUTTER_TEST');